Refactor to browser-native ES modules: extract utils, PWA registration, SVG and tooltip helpers; switch index.html to module script; update SW cache to include new modules
