+++ index.html
@@ -122,6 +122,6 @@
   <!-- Touch ability popup -->
   <div id="touch-ability-popup" class="touch-ability-popup"></div>
 
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
